date	2005.04.22.14.33.00;	author endre;	state Exp;
git-svn-id: https://svn.isis.vanderbilt.edu/MoBIES/UDM/trunk@2164 c0adbae9-c137-0410-b742-ec4b2d2ceded

+++ src/UdmOclPat/GOCL/src/OCLObjectExBasic.cpp
@@ -293,7 +293,7 @@ namespace OclBasic
 
 	void Integer::SetValue( long value )
 	{
-		SetValue(  value );
+		SetValue( (double) value );
 	}
 
 	Any* Integer::Clone() const
